pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "cython>=0.29.0"]
build-backend = "setuptools.build_meta"

[project]
name = "flashback"
version = "0.1.0"
description = "High-Frequency Trading Strategy Backtester"
authors = [{name = "Flashback Team"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "cython>=0.29.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
flashback = "flashback.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["flashback*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

